requirements.txt
pandas
ta
python-dotenv
binance-connector 